requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
distlib==0.3.1
filelock==3.0.12
googlesearch-python==1.0.1
idna==2.10
lxml==4.6.3
pipenv==2021.5.29
pytube==11.0.0
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
typing-extensions==3.10.0.0
urllib3==1.26.6
virtualenv==20.3.1
virtualenv-clone==0.5.6
youtube-dl==2021.6.6